+++ bankify.py
@@ -1,176 +1,9 @@
 from __future__ import annotations
 from pathlib import Path
-from typing import Dict, Iterable, Optional, Tuple, Union
-from zipfile import ZipFile, ZIP_DEFLATED
+from typing import Dict, Optional, Tuple
+from vital import Bank, VitalObject
 import json
 
-
-class Lfo:
-    """
-    A Vital LFO
-    """
-    EXTENSION = 'vitallfo'
-    DIR = 'LFOs'
-
-    @classmethod
-    def from_file(cls, filepath: Path) -> Lfo:
-        """
-        Create an LFO from a file
-        :param filepath:
-        :return:
-        """
-        with filepath.open('r') as f:
-            return cls(data=json.load(f), filepath=filepath)
-
-    def __str__(self):
-        return self.name
-
-    def __repr__(self):
-        return str(self)
-
-    def __init__(self, data: dict, filepath: Optional[Path]):
-        self._data = data
-        self.original_path = filepath
-        self.name = self._data['name']
-
-    def rename(self, new_name: str):
-        """
-        Rename
-
-        :param new_name:
-        :return:
-        """
-        self._data['name'] = new_name
-        self.name = self._data['name']
-
-
-class Preset:
-    """
-    A Vital Preset
-    """
-    EXTENSION = 'vital'
-    DIR = 'Presets'
-
-    @classmethod
-    def from_file(cls, filepath: Path) -> Preset:
-        """
-        Create an Preset from a file
-        :param filepath:
-        :return:
-        """
-        with filepath.open('r') as f:
-            return cls(data=json.load(f), filepath=filepath)
-
-    def __str__(self):
-        return self.name
-
-    def __repr__(self):
-        return str(self)
-
-    def __init__(self, data: dict, filepath: Optional[Path], **kwargs):
-        self._data = data
-        self.original_path = filepath
-
-        if filepath is not None:  # Differs  because it appears the name comes from the file
-            self.name = filepath.stem
-        else:
-            self.name = kwargs['name']
-
-    def rename(self, new_name: str):
-        """
-        Rename
-
-        :param new_name:
-        :return:
-        """
-        self.name = new_name
-
-
-class Wavetable:
-    """
-    A Vital Wavetable
-    """
-    EXTENSION = 'vitaltable'
-    DIR = 'Wavetables'
-
-    @classmethod
-    def from_file(cls, filepath: Path) -> Wavetable:
-        """
-        Create an LFO from a file
-        :param filepath:
-        :return:
-        """
-        with filepath.open('r') as f:
-            return cls(data=json.load(f), filepath=filepath)
-
-    def __str__(self):
-        return self.name
-
-    def __repr__(self):
-        return str(self)
-
-    def __init__(self, data: dict, filepath: Optional[Path]):
-        self._data = data
-        self.original_path = filepath
-        self.name = self._data['name']
-
-    def rename(self, new_name: str):
-        """
-        Rename
-
-        :param new_name:
-        :return:
-        """
-        self._data['name'] = new_name
-        self.name = self._data['name']
-
-
-class Bank:
-    """
-    Represents a bank comprised of LFOs, Presets, and Wavetables
-    """
-    EXTENSION = 'vitalbank'
-
-    def __str__(self):
-        return self.name
-
-    def __repr__(self):
-        return str(self)
-
-    def __init__(self, bank_name: str, lfos: Optional[Iterable[Lfo]], presets: Optional[Iterable[Preset]],
-                 wavetables: Optional[Iterable[Wavetable]]):
-        self.name = bank_name
-        self.elements = {
-            Lfo: {lfo.name: lfo for lfo in lfos} if lfos is not None else list(),
-            Preset: {preset.name: preset for preset in presets} if presets is not None else list(),
-            Wavetable: {wavetable.name: wavetable for wavetable in wavetables} if wavetables is not None else list(),
-        }
-
-    def write_file(self, filedir: Path):
-        """
-        Write the bank file to disk
-
-        :param filedir: The directory to write to
-        :return:
-        """
-        filepath = filedir.joinpath(f'{self.name}.{Bank.EXTENSION}')
-
-        bankfile = ZipFile(file=filepath, mode='w', compression=ZIP_DEFLATED, compresslevel=9)
-
-        for ct, ctd in self.elements.items():
-            zdir = f'{self.name}/{ct.DIR}'
-            for i in ctd:
-                zfile = f'{zdir}/{i.name}.{ct.EXTENSION}'
-                bankfile.writestr(zinfo_or_arcname=zfile, data=json.dumps(i._data))
-
-        bankfile.close()
-
-
-TYPE_DATA = {
-    Lfo: ('LFOs', 'vitallfo'),
-    Preset: ('Presets', 'vital'),
-    Wavetable: ('Wavetables', 'vitaltable'),
-}
 
 CONFIG_FILE = Path('config.json')
 with CONFIG_FILE.open('r') as config_file:
@@ -181,7 +14,7 @@ with CONFIG_FILE.open('r') as config_file:
     BANK_DELIM = CONFIG['delimiter']
 
 
-def is_bank_obj(delim: str, vobj: Union[Lfo, Preset, Wavetable]) -> Optional[Union[Lfo, Preset, Wavetable]]:
+def is_bank_obj(delim: str, vobj: VitalObject) -> Optional[VitalObject]:
     """
     If an object name contains the delmiter and matches the pattern for belonging in a bank, return it. None otherwise.
     First delimiter must be at the beginning
@@ -200,7 +33,7 @@ def is_bank_obj(delim: str, vobj: Union[Lfo, Preset, Wavetable]) -> Optional[Uni
         return None
 
 
-def bank_comps(delim: str, vobj: Union[Lfo, Preset, Wavetable]) -> Tuple[str, str]:
+def bank_comps(delim: str, vobj: VitalObject) -> Tuple[str, str]:
     """
     Return the bank name and the name of the object with the bank bits stripped out.
 
@@ -223,16 +56,16 @@ def bank_comps(delim: str, vobj: Union[Lfo, Preset, Wavetable]) -> Tuple[str, st
 def main():
     banks: Dict[str, Bank] = dict()
 
-    for clstype, typedata in TYPE_DATA.items():
-        subdir = USER_DIR.joinpath(typedata[0])
+    for clstype in VitalObject.__subclasses__():
+        subdir = USER_DIR.joinpath(clstype.DIR)
         for f in subdir.glob('*'):
             if not f.is_dir() and f.suffix == f'.{clstype.EXTENSION}':
                 vobj = clstype.from_file(filepath=f)
                 if is_bank_obj(delim=BANK_DELIM, vobj=vobj) is not None:
                     bank_name, obj_name = bank_comps(delim=BANK_DELIM, vobj=vobj)
-                    banks.setdefault(bank_name, Bank(bank_name=bank_name, lfos=None, presets=None, wavetables=None))
+                    banks.setdefault(bank_name, Bank(bank_name=bank_name, vital_objects=None))
                     vobj.rename(new_name=obj_name)  # Set the "clean" name
-                    banks[bank_name].elements[clstype].append(vobj)
+                    banks[bank_name].add_object(vital_object=vobj)
 
     for bank in banks.values():
         print(bank.name)